added Dockerfile and docs
